Start rework
I added the fields that I think that I will need to basic blocks and
made the arguments to jump functions exist properly. So that will help
with the later development.

While I am focusing on #12 foremost, there is always the idea of the
other things in the background. I think that I'm going to refactor all
the functions to have trailing types that way when you're writing out
the function it has the consistent naming convention across different
things. Predominantly this is of the form link_bb instead of bb_link.
But I also changed it from add_bb to create_bb because it is making the
bb and not adding an existing thing to the list of bbs. I think that at
some level this makes more sense.

I really want to clean up the code, it reached a really messy state...