Need to fix structure
Was going to really quick go through and implement the skeleton
generation for the cfg before going to bed. I had yet to think through
the cases where we are going to have to do each expression expansion. I
think that each statement should only produce a singular basic block.

As increasing layers of internal recursion for macros are just going to
push the requirements up to the block above them to then be parsed into
this block.

In this manner I think possibly a good idea would be to move around the
statements or on the fly generate more statements before doing the
instruction generation as this would allow us to split up the dependent
macro case which is the entire reason for why I'm doing this rewrite of
the exisitng cfg generation.